backend/requirements.txt
Flask
numpy
torch
pyyaml
librosa
